refactor: move compact binary tests in new file
